@@ -4,17 +4,21 @@
   "private": true,
   "homepage": "https://dehnes.com/smarthome",
   "dependencies": {
-    "@material-ui/core": "^4.12.4",
-    "@material-ui/icons": "^4.11.3",
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
+    "@mdi/js": "^7.2.96",
+    "@mdi/react": "^1.6.1",
+    "@mui/icons-material": "^5.11.16",
+    "@mui/material": "^5.12.1",
     "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.3.0",
+    "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.2.1",
-    "@types/jest": "^28.1.4",
+    "@types/jest": "^29.5.1",
     "@types/node": "^18.0.0",
     "@types/react": "^18.0.14",
     "@types/react-dom": "^18.0.5",
     "@types/react-router-dom": "^5.3.3",
-    "@types/uuid": "^8.3.4",
+    "@types/uuid": "^9.0.1",
     "chart.js": "^4.0.1",
     "class-transformer": "^0.5.1",
     "react": "^18.2.0",
@@ -23,9 +27,9 @@
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
     "reflect-metadata": "^0.1.13",
-    "typescript": "^4.7.4",
-    "uuid": "^8.3.2",
-    "web-vitals": "^2.1.4"
+    "typescript": "^4.9.5",
+    "uuid": "^9.0.0",
+    "web-vitals": "^3.3.1"
   },
   "scripts": {
     "start": "react-scripts start",
