@@ -18,8 +18,8 @@ import {
     TableCell,
     TableContainer,
     TableRow
-} from "@material-ui/core";
-import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
+} from "@mui/material";
+import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
 import { timeToDelta } from "../GarageDoor/GarageDoor";
 import { useUserSettings } from "../../Websocket/websocketClient";
 
@@ -109,190 +109,192 @@ export const EnvironmentSensor = ({
         sleepTimeInSecondsDelta: delta
     });
 
-    return <Accordion key={sensor.sensorId}>
-        <AccordionSummary
-            expandIcon={<ExpandMoreIcon/>}
-            aria-controls="panel1a-content"
-            id="panel1a-header"
-        >
-            <div style={{
-                display: "flex",
-                flexDirection: "row",
-                alignItems: "center",
-                width: "100%",
-                justifyContent: "space-between"
-            }}>
+    return (
+        <Accordion key={sensor.sensorId}>
+            <AccordionSummary
+                expandIcon={<ExpandMoreIcon/>}
+                aria-controls="panel1a-content"
+                id="panel1a-header"
+            >
                 <div style={{
-                    fontWeight: "bold",
-                    fontSize: "140%"
-                }}>{sensor.displayName}</div>
-                <div>
-                    {stateToText(currentMilliSeconds, sensor)}
-                    {statusIcon(sensor, currentMilliSeconds)}
+                    display: "flex",
+                    flexDirection: "row",
+                    alignItems: "center",
+                    width: "100%",
+                    justifyContent: "space-between"
+                }}>
+                    <div style={{
+                        fontWeight: "bold",
+                        fontSize: "140%"
+                    }}>{sensor.displayName}</div>
+                    <div>
+                        {stateToText(currentMilliSeconds, sensor)}
+                        {statusIcon(sensor, currentMilliSeconds)}
+                    </div>
                 </div>
-            </div>
-        </AccordionSummary>
-        <AccordionDetails>
+            </AccordionSummary>
+            <AccordionDetails>
 
-            <Grid container spacing={2}>
-                <Grid item xs={12}>
-                    <Grid container justify="space-between" spacing={2}>
-                        <ButtonGroup variant="contained" style={{
-                            margin: "10px"
-                        }}>
-                            <Button
-                                disabled={!userSettings.userCanAdmin("environmentSensors")}
-                                color={sensor.timeAdjustmentSchedule ? 'secondary' : 'primary'}
-                                onClick={() =>
-                                    sensor.timeAdjustmentSchedule
-                                        ? sendCommand("cancelTimeAdjustment")
-                                        : sendCommand("scheduleTimeAdjustment")
-                                }
-                            >Adjust time</Button>
-                            <Button
-                                disabled={!firmwareInfo?.filename || !userSettings.userCanAdmin("environmentSensors")}
-                                color={sensor.firmwareUpgradeScheduled ? 'secondary' : 'primary'}
-                                onClick={() =>
-                                    sensor.firmwareUpgradeScheduled
-                                        ? sendCommand("cancelFirmwareUpgrade")
-                                        : sendCommand("scheduleFirmwareUpgrade")
-                                }
-                            >Upgrade firmware</Button>
-                            <Button
-                                disabled={!userSettings.userCanAdmin("environmentSensors")}
-                                color={sensor.resetScheduled ? 'secondary' : 'primary'}
-                                onClick={() =>
-                                    sensor.resetScheduled
-                                        ? sendCommand("cancelReset")
-                                        : sendCommand("scheduleReset")
-                                }
-                            >Reset</Button>
-                        </ButtonGroup>
+                <Grid container spacing={2}>
+                    <Grid item xs={12}>
+                        <Grid container justifyContent="space-between" spacing={2}>
+                            <ButtonGroup variant="contained" style={{
+                                margin: "10px"
+                            }}>
+                                <Button
+                                    disabled={!userSettings.userCanAdmin("environmentSensors")}
+                                    color={sensor.timeAdjustmentSchedule ? 'secondary' : 'primary'}
+                                    onClick={() =>
+                                        sensor.timeAdjustmentSchedule
+                                            ? sendCommand("cancelTimeAdjustment")
+                                            : sendCommand("scheduleTimeAdjustment")
+                                    }
+                                >Adjust time</Button>
+                                <Button
+                                    disabled={!firmwareInfo?.filename || !userSettings.userCanAdmin("environmentSensors")}
+                                    color={sensor.firmwareUpgradeScheduled ? 'secondary' : 'primary'}
+                                    onClick={() =>
+                                        sensor.firmwareUpgradeScheduled
+                                            ? sendCommand("cancelFirmwareUpgrade")
+                                            : sendCommand("scheduleFirmwareUpgrade")
+                                    }
+                                >Upgrade firmware</Button>
+                                <Button
+                                    disabled={!userSettings.userCanAdmin("environmentSensors")}
+                                    color={sensor.resetScheduled ? 'secondary' : 'primary'}
+                                    onClick={() =>
+                                        sensor.resetScheduled
+                                            ? sendCommand("cancelReset")
+                                            : sendCommand("scheduleReset")
+                                    }
+                                >Reset</Button>
+                            </ButtonGroup>
+                        </Grid>
                     </Grid>
-                </Grid>
-                {sensor.sensorData &&
-                    <>
-                        <Grid item xs={12}>
-                            <Grid container justify="flex-start" spacing={2} alignItems={"center"}>
+                    {sensor.sensorData &&
+                        <>
+                            <Grid item xs={12}>
+                                <Grid container justifyContent="flex-start" spacing={2} alignItems={"center"}>
 
-                                <Grid item xs={8}>
-                                    <span>Set sleep time: </span>
-                                    <span>{sensor.sleepTimeInSeconds} seconds</span>
-                                </Grid>
-                                <Grid item xs={4}>
-                                    <ButtonGroup variant="contained" aria-label="contained primary button group"
-                                                 style={{
-                                                     margin: "10px"
-                                                 }}>
-                                        <Button
-                                            disabled={sensor.sleepTimeInSeconds < 11 || !userSettings.userCanAdmin("environmentSensors")}
-                                            onClick={() => adjustSleepTimeInSeconds(-10)}>-</Button>
-                                        <Button
-                                            disabled={sensor.sleepTimeInSeconds >= 590 || !userSettings.userCanAdmin("environmentSensors")}
-                                            onClick={() => adjustSleepTimeInSeconds(10)}>+</Button>
-                                    </ButtonGroup>
+                                    <Grid item xs={8}>
+                                        <span>Set sleep time: </span>
+                                        <span>{sensor.sleepTimeInSeconds} seconds</span>
+                                    </Grid>
+                                    <Grid item xs={4}>
+                                        <ButtonGroup variant="contained" aria-label="contained primary button group"
+                                                     style={{
+                                                         margin: "10px"
+                                                     }}>
+                                            <Button
+                                                disabled={sensor.sleepTimeInSeconds < 11 || !userSettings.userCanAdmin("environmentSensors")}
+                                                onClick={() => adjustSleepTimeInSeconds(-10)}>-</Button>
+                                            <Button
+                                                disabled={sensor.sleepTimeInSeconds >= 590 || !userSettings.userCanAdmin("environmentSensors")}
+                                                onClick={() => adjustSleepTimeInSeconds(10)}>+</Button>
+                                        </ButtonGroup>
+                                    </Grid>
                                 </Grid>
                             </Grid>
-                        </Grid>
 
+                            <Grid item xs={12}>
+                                <Grid container justifyContent="flex-start" spacing={2} alignItems={"center"}>
+                                    <TableContainer component={Paper} style={{
+                                        marginTop: "20px"
+                                    }}>
+                                        <Table aria-label="simple table">
+                                            <TableBody>
+                                                <TableRow>
+                                                    <TableCell component="th" scope="row">Temperature</TableCell>
+                                                    <TableCell
+                                                        align="right">{(sensor.sensorData!!.temperature / 100).toFixed(2)} &deg;C</TableCell>
+                                                </TableRow>
+                                                <TableRow>
+                                                    <TableCell component="th" scope="row">Temperature Error</TableCell>
+                                                    <TableCell
+                                                        align="right">{sensor.sensorData.temperatureError ? 'true' : 'false'}</TableCell>
+                                                </TableRow>
+                                                <TableRow>
+                                                    <TableCell component="th" scope="row">Humidity</TableCell>
+                                                    <TableCell
+                                                        align="right">{(sensor.sensorData!!.humidity / 100).toFixed(2)} %</TableCell>
+                                                </TableRow>
+                                                <TableRow>
+                                                    <TableCell component="th" scope="row">Light</TableCell>
+                                                    <TableCell align="right">{sensor.sensorData!!.adcLight}</TableCell>
+                                                </TableRow>
+                                                <TableRow>
+                                                    <TableCell component="th" scope="row">Battery</TableCell>
+                                                    <TableCell
+                                                        align="right">{(sensor.sensorData!!.batteryMilliVolts / 1000).toFixed(2)} V</TableCell>
+                                                </TableRow>
+                                                <TableRow>
+                                                    <TableCell component="th" scope="row">Actual sleep time</TableCell>
+                                                    <TableCell
+                                                        align="right">{sensor.sensorData!!.sleepTimeInSeconds} seconds</TableCell>
+                                                </TableRow>
+                                                <TableRow>
+                                                    <TableCell component="th" scope="row">Firmware version</TableCell>
+                                                    <TableCell align="right">{sensor.firmwareVersion}</TableCell>
+                                                </TableRow>
+                                                <TableRow>
+                                                    <TableCell component="th" scope="row">Clock slew</TableCell>
+                                                    <TableCell
+                                                        align="right">{sensor.sensorData!!.timestampDelta} seconds</TableCell>
+                                                </TableRow>
+                                                <TableRow>
+                                                    <TableCell component="th" scope="row">Received:</TableCell>
+                                                    <TableCell
+                                                        align="right">{timeToDelta(currentMilliSeconds, sensor.sensorData!!.receivedAt)} ago</TableCell>
+                                                </TableRow>
+                                                <TableRow>
+                                                    <TableCell component="th" scope="row">Rssi:</TableCell>
+                                                    <TableCell
+                                                        align="right">{sensor.sensorData!!.rssi}dB</TableCell>
+                                                </TableRow>
+                                            </TableBody>
+                                        </Table>
+                                    </TableContainer>
+                                </Grid>
+                            </Grid>
+                        </>
+                    }
+
+                    {sensor.firmwareUpgradeState &&
                         <Grid item xs={12}>
-                            <Grid container justify="flex-start" spacing={2} alignItems={"center"}>
+                            <Grid container justifyContent="flex-start" spacing={2} alignItems={"center"}>
                                 <TableContainer component={Paper} style={{
                                     marginTop: "20px"
                                 }}>
                                     <Table aria-label="simple table">
                                         <TableBody>
                                             <TableRow>
-                                                <TableCell component="th" scope="row">Temperature</TableCell>
-                                                <TableCell
-                                                    align="right">{(sensor.sensorData!!.temperature / 100).toFixed(2)} &deg;C</TableCell>
-                                            </TableRow>
-                                            <TableRow>
-                                                <TableCell component="th" scope="row">Temperature Error</TableCell>
-                                                <TableCell
-                                                    align="right">{sensor.sensorData.temperatureError ? 'true' : 'false'}</TableCell>
-                                            </TableRow>
-                                            <TableRow>
-                                                <TableCell component="th" scope="row">Humidity</TableCell>
-                                                <TableCell
-                                                    align="right">{(sensor.sensorData!!.humidity / 100).toFixed(2)} %</TableCell>
-                                            </TableRow>
-                                            <TableRow>
-                                                <TableCell component="th" scope="row">Light</TableCell>
-                                                <TableCell align="right">{sensor.sensorData!!.adcLight}</TableCell>
-                                            </TableRow>
-                                            <TableRow>
-                                                <TableCell component="th" scope="row">Battery</TableCell>
-                                                <TableCell
-                                                    align="right">{(sensor.sensorData!!.batteryMilliVolts / 1000).toFixed(2)} V</TableCell>
-                                            </TableRow>
-                                            <TableRow>
-                                                <TableCell component="th" scope="row">Actual sleep time</TableCell>
-                                                <TableCell
-                                                    align="right">{sensor.sensorData!!.sleepTimeInSeconds} seconds</TableCell>
-                                            </TableRow>
-                                            <TableRow>
-                                                <TableCell component="th" scope="row">Firmware version</TableCell>
-                                                <TableCell align="right">{sensor.firmwareVersion}</TableCell>
+                                                <TableCell component="th" scope="row">Progress:</TableCell>
+                                                <TableCell align="right">{firmwareUpgradeProgress(sensor)}%</TableCell>
                                             </TableRow>
                                             <TableRow>
                                                 <TableCell component="th" scope="row">Clock slew</TableCell>
                                                 <TableCell
-                                                    align="right">{sensor.sensorData!!.timestampDelta} seconds</TableCell>
+                                                    align="right">{sensor.firmwareUpgradeState!!.timestampDelta} seconds</TableCell>
                                             </TableRow>
                                             <TableRow>
                                                 <TableCell component="th" scope="row">Received:</TableCell>
                                                 <TableCell
-                                                    align="right">{timeToDelta(currentMilliSeconds, sensor.sensorData!!.receivedAt)} ago</TableCell>
+                                                    align="right">{timeToDelta(currentMilliSeconds, sensor.firmwareUpgradeState!!.receivedAt)} ago</TableCell>
                                             </TableRow>
                                             <TableRow>
                                                 <TableCell component="th" scope="row">Rssi:</TableCell>
                                                 <TableCell
-                                                    align="right">{sensor.sensorData!!.rssi}dB</TableCell>
+                                                    align="right">{sensor.firmwareUpgradeState!!.rssi}dB</TableCell>
                                             </TableRow>
                                         </TableBody>
                                     </Table>
                                 </TableContainer>
                             </Grid>
                         </Grid>
-                    </>
-                }
+                    }
 
-                {sensor.firmwareUpgradeState &&
-                    <Grid item xs={12}>
-                        <Grid container justify="flex-start" spacing={2} alignItems={"center"}>
-                            <TableContainer component={Paper} style={{
-                                marginTop: "20px"
-                            }}>
-                                <Table aria-label="simple table">
-                                    <TableBody>
-                                        <TableRow>
-                                            <TableCell component="th" scope="row">Progress:</TableCell>
-                                            <TableCell align="right">{firmwareUpgradeProgress(sensor)}%</TableCell>
-                                        </TableRow>
-                                        <TableRow>
-                                            <TableCell component="th" scope="row">Clock slew</TableCell>
-                                            <TableCell
-                                                align="right">{sensor.firmwareUpgradeState!!.timestampDelta} seconds</TableCell>
-                                        </TableRow>
-                                        <TableRow>
-                                            <TableCell component="th" scope="row">Received:</TableCell>
-                                            <TableCell
-                                                align="right">{timeToDelta(currentMilliSeconds, sensor.firmwareUpgradeState!!.receivedAt)} ago</TableCell>
-                                        </TableRow>
-                                        <TableRow>
-                                            <TableCell component="th" scope="row">Rssi:</TableCell>
-                                            <TableCell
-                                                align="right">{sensor.firmwareUpgradeState!!.rssi}dB</TableCell>
-                                        </TableRow>
-                                    </TableBody>
-                                </Table>
-                            </TableContainer>
-                        </Grid>
-                    </Grid>
-                }
-
-            </Grid>
-        </AccordionDetails>
-    </Accordion>;
+                </Grid>
+            </AccordionDetails>
+        </Accordion>
+    );
 };
