@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import WebsocketService from "../../Websocket/websocketClient";
-import { Container } from "@material-ui/core";
+import { Container } from "@mui/material";
 import Header from "../Header";
 import {
     EnvironmentSensorRequest,
