@@ -1,8 +1,17 @@
-import {createTheme} from '@material-ui/core/styles';
+import {createTheme} from '@mui/material/styles';
 // A custom theme for this app
 const theme = createTheme({
     palette: {
-        type: 'dark',
-    }
+        mode: 'dark',
+        primary: {
+            main: '#2457ff',
+            contrastText: '#ffffff',
+        },
+        secondary: {
+            main: '#ff006a',
+            contrastText: '#ffffff',
+        }
+    },
 });
+
 export default theme;
